+++ src/constants/selectCurrency.config.tsx
@@ -1,7 +1,0 @@
-import {StyleSheet} from 'react-native'
-import {Icon} from '@rneui/themed'
-
-import type {PlatformType} from 'hooks/helper/usePlatform'
-
-
-export const availableCurrency= ['USD', ]