+++ src/screens/BuyToken/CurrencySelector/CurrencySelector.tsx
@@ -1,21 +1,19 @@
 import React from 'react'
-import {View} from 'react-native'
-import {Button, Text} from '@rneui/themed'
+import {View, TouchableOpacity} from 'react-native'
+import {Divider, Text} from '@rneui/themed'
 
 import BottomSheet, {BottomSheetProps} from '@core/BottomSheet'
 
 import {usePlatform} from 'hooks/helper'
-import {getCurrencyConfig} from 'constants/currency.config'
+import {AllCurrencyType} from 'constants/currency.config'
 
-// I'm still working on currency selector
+import {getCurrencyConfig} from '../currencySelector.config'
 
-const CurrencySelect: React.FC<Omit<BottomSheetProps, 'children' | 'title'>> = ({
-  isOpened,
-  onClose,
-}) => {
+const CurrencySelect: React.FC<
+  Omit<BottomSheetProps, 'children' | 'title'> & {onPress: (id: AllCurrencyType) => void}
+> = ({isOpened, onClose, onPress}) => {
   const {platform} = usePlatform()
   const currencyConfig = getCurrencyConfig(platform)
-
 
   return (
     <BottomSheet
@@ -26,15 +24,23 @@ const CurrencySelect: React.FC<Omit<BottomSheetProps, 'children' | 'title'>> = (
         rowGap: 15,
       }}
     >
-      {currencyConfig.map(({id, label, icon: Icon}) => (
-        <View key={id}>
-          <View style={{flexDirection: 'row', alignItems: 'center', columnGap: 10}}>
-            {Icon}
-            <Text style={{fontSize: 16, fontWeight: '600'}}>{label}</Text>
-          </View>
+      {currencyConfig.map((item, index, array) => (
+        <View key={item.id}>
+          <TouchableOpacity
+            onPress={() => {
+              onClose()
+              onPress(item.id)
+            }}
+          >
+            <View style={{flexDirection: 'row', alignItems: 'center', columnGap: 10}}>
+              {item.icon}
+              <Text style={{fontSize: 16, fontWeight: '700'}}>{item.label}</Text>
+            </View>
+          </TouchableOpacity>
+
+          {index + 1 !== array.length && <Divider style={{marginTop: 10}} />}
         </View>
       ))}
-      <Button title='change currency' />
     </BottomSheet>
   )
 }