+++ RELEASE_NOTES.md
@@ -1,5 +1,15 @@
 # RELEASE NOTES
 
+## Released APP v0.1.1
+
+### Added
+
+- Enabled currency switching
+
+### Changed
+
+- Modified tier system modal
+
 ## Platform APP v0.1.0 - 2023-11-15
 
 ### Added